requirements.txt
adlfs==2022.11.2
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anthropic==0.3.11
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
astroid==3.0.3
asttokens==2.2.1
async-lru==2.0.3
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.3
azure-datalake-store==0.0.53
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.17.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.12.0
bleach==6.0.0
boto3==1.28.49
botocore==1.31.49
cachetools==5.3.1
certifi==2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5567-1hyue8f/cffi-1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.5
cohere==4.26.1
colorama==0.4.6
comm==0.1.3
contextlib2==21.6.0
contourpy==1.1.0
courlan==0.9.4
cryptography==41.0.3
cycler==0.11.0
dacite==1.8.1
dataclasses-json==0.5.12
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
delta==0.4.2
delta-spark==3.0.0
deltalake==0.14.0
diff-cover==7.7.0
dill==0.3.8
distlib @ file:///private/tmp/python-distlib-20231124-5231-2n0vb3/distlib-0.3.7
distro==1.8.0
dnspython==2.4.0
docker==5.0.3
docutils @ file:///private/tmp/docutils-20231015-5454-w1xze4/docutils-0.20.1
executing==1.2.0
fastapi==0.100.0
fastavro==1.8.2
fastjsonschema==2.17.1
filelock==3.12.3
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-aiplatform==1.34.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
GQLAlchemy==1.5.0
griptape==0.17.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
htmldate==1.5.1
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
joblibspark==0.5.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
jusText==3.0.0
kiwisolver==1.4.4
langchain==0.0.323
langcodes==3.3.0
langsmith==0.0.52
litellm==0.13.2
livy==0.8.0
loguru==0.7.2
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marqo==1.3.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
neo4j==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opensearch-py==2.3.1
oscrypto==1.3.0
overrides==7.3.1
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.4
platformdirs==3.7.0
pluggy==1.3.0
ply==3.11
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pydantic-settings==2.0.2
pydantic_core==2.10.1
Pygments==2.15.1
PyJWT==2.8.0
pylint==3.0.3
pymgclient==1.3.1
pymongo==4.4.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
PyPDF2==3.0.1
pyspark==3.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.6.0
referencing==0.29.1
regex==2023.8.8
requests==2.31.0
requests-aws4auth==1.2.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.8.10
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
schema==0.7.5
Send2Trash==1.8.2
Shapely==1.8.5.post1
simple-ddl-parser==0.31.0
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.2.0
snowflake-sqlalchemy==1.5.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlalchemy-redshift==0.8.14
sqlfluff==2.1.4
sqllineage==1.4.7
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
stopit==1.1.2
stringcase==1.2.0
tabulate==0.9.0
tblib==2.0.0
tenacity==8.2.2
terminado==0.17.1
testbook==0.4.2
tiktoken==0.5.1
tinycss2==1.2.1
tld==0.13
tokenizers==0.13.3
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
trafilatura==1.6.2
traitlets==5.9.0
transformers==4.33.2
truststore==0.8.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.1
vertexai==0.0.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.8
XlsxWriter==3.2.0
yarl==1.9.2
zipp==3.17.0
